+++ decision_tree_src/image_availability.py
